Deep copy instead of shallow copy for active layer
dxf_set_layer previously stored a pointer to the active layer and
assumed that it remained valid for as long as it might be used. That was
fine when all layer names were constant, but it's more of a pain with
custom layer names that have shorter life cycles. This removes the
footgun.